@@ -1,11 +1,8 @@
-import Vue from 'vue'
-import Vuex from 'vuex'
-import Quiz from './modules/quiz'
+import { createStore } from 'vuex';
+import Quiz from './modules/quiz';
 
-Vue.use(Vuex)
-
-export default new Vuex.Store({
+export default createStore({
   modules: {
     Quiz
   }
-})
+});
